Move cli into common pkg and cmd directory structure
This is the third commit in the chain for reorganising under a common
cmd and pkg directory structure. This commit moves the cli under the cmd
and pkg directory structure.
