+++ pkg/cli/cli.go
@@ -25,8 +25,8 @@ import (
 
 	"github.com/google/uuid"
 
-	"github.com/openclarity/vmclarity/cli/pkg/presenter"
-	"github.com/openclarity/vmclarity/cli/pkg/state"
+	"github.com/openclarity/vmclarity/pkg/cli/presenter"
+	"github.com/openclarity/vmclarity/pkg/cli/state"
 	"github.com/openclarity/vmclarity/pkg/shared/families"
 	"github.com/openclarity/vmclarity/pkg/shared/families/types"
 	"github.com/openclarity/vmclarity/pkg/shared/fsutils/blockdevice"
