+++ pkg/cli/version.go
@@ -13,12 +13,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package cli
 
-import (
-	"github.com/openclarity/vmclarity/cli/cmd"
-)
-
-func main() {
-	cmd.Execute()
-}
+var GitRevision string
